+++ src/ramifice/utils/mixins/add_valid.py
@@ -4,10 +4,12 @@ __all__ = ("AddValidMixin",)
 
 from abc import ABCMeta
 
+from xloft import NamedTuple
+
 
 class AddValidMixin(metaclass=ABCMeta):
     """Ramifice - Contains an abstract method for additional validation of fields."""
 
-    async def add_validation(self) -> dict[str, str]:
+    async def add_validation(self) -> NamedTuple:
         """Ramifice - Additional validation of fields."""
-        return {}
+        return NamedTuple()